+++ examples/border-test.rs
@@ -28,12 +28,16 @@ macro_rules! result { ($type: ty) => { Result<$type, NCurseswWinError> } }
 
 fn main() {
     // initialize ncurses in a safe way.
-    if let Err(source) = ncursesw_entry(|window| {
+    if let Err(source) = ncursesw_entry(|stdscr| {
+        set_input_mode(InputMode::Character)?;
+        set_echo(false)?;
+        set_newline(false)?;
+        intrflush(false)?;
+
         // set the cursor to invisible and switch echoing off.
         cursor_set(CursorType::Invisible)?;
-        set_echo(false)?;
 
-        border_test(window)
+        border_test(stdscr)
     }) {
         match source {
             NCurseswWinError::Panic { message } => eprintln!("panic: {}", message),